+++ waterfall-cad/src/Waterfall/Booleans.hs
@@ -6,6 +6,8 @@ module Waterfall.Booleans
 , difference
 , intersection
 , complement
+, unions
+, intersections
 ) where
 
-import Waterfall.Internal.Solid(union, difference, intersection, complement)
+import Waterfall.Internal.Solid(union, unions, difference, intersection, intersections, complement)